fix(amazonq): Bug fixes for LSP auth on agentic mode (#7231)
## Problem
Agentic chat introduced code changes that break auth on LSP

## Solution
Fixes for the auth flow:
* Move instantiation of `AuthUtil` into `startLanguageServer` sinc ethe
function uses several AuthUtil calls, which fail if it is not
instantiated yet
* Remove the dependency of LSP on glibc patch, since auth doesn't work
if LSP is not started
* Add URL opening capability to the new `ShowDocumentRequest` handler,
and remove the old handler from `client.ts`.
* Fix bug where session isn't restored upon start of extension when SSO
is expired.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.
