+++ packages/amazonq/src/extensionNode.ts
@@ -19,7 +19,6 @@ import api from './api'
 import { activate as activateCWChat } from './app/chat/activation'
 import { activate as activateInlineChat } from './inlineChat/activation'
 import { beta } from 'aws-core-vscode/dev'
-import * as amazonq from 'aws-core-vscode/amazonq'
 import { activate as activateNotifications, NotificationsController } from 'aws-core-vscode/notifications'
 import { AuthUtil } from 'aws-core-vscode/codewhisperer'
 import { telemetry, AuthUserState } from 'aws-core-vscode/telemetry'
@@ -67,8 +66,6 @@ async function activateAmazonQNode(context: vscode.ExtensionContext) {
     activateAgents()
     await activateTransformationHub(extContext as ExtContext)
     activateInlineChat(context)
-
-    context.subscriptions.push(amazonq.focusAmazonQPanel.register(), amazonq.focusAmazonQPanelKeybinding.register())
 
     const authProvider = new CommonAuthViewProvider(
         context,
