+++ packages/amazonq/src/lsp/activation.ts
@@ -4,19 +4,17 @@
  */
 
 import vscode from 'vscode'
-import { clientId, encryptionKey, startLanguageServer } from './client'
+import { startLanguageServer } from './client'
 import { AmazonQLspInstaller } from './lspInstaller'
 import { lspSetupStage, ToolkitError, messages } from 'aws-core-vscode/shared'
 import { AuthUtil } from 'aws-core-vscode/codewhisperer'
-import { auth2 } from 'aws-core-vscode/auth'
 
 export async function activate(ctx: vscode.ExtensionContext) {
     try {
-        const client = await lspSetupStage('all', async () => {
+        await lspSetupStage('all', async () => {
             const installResult = await new AmazonQLspInstaller().resolve()
             return await lspSetupStage('launch', () => startLanguageServer(ctx, installResult.resourcePaths))
         })
-        AuthUtil.create(new auth2.LanguageClientAuth(client, clientId, encryptionKey))
         await AuthUtil.instance.restore()
     } catch (err) {
         const e = err as ToolkitError
