+++ packages/amazonq/src/lsp/chat/messages.ts
@@ -36,9 +36,6 @@ import {
     ShowSaveFileDialogParams,
     LSPErrorCodes,
     tabBarActionRequestType,
-    ShowDocumentParams,
-    ShowDocumentResult,
-    ShowDocumentRequest,
     contextCommandsNotificationType,
     ContextCommandParams,
     openFileDiffNotificationType,
@@ -422,23 +419,6 @@ export function registerMessageListeners(
             targetUri: targetUri.toString(),
         }
     })
-
-    languageClient.onRequest<ShowDocumentParams, ShowDocumentResult>(
-        ShowDocumentRequest.method,
-        async (params: ShowDocumentParams): Promise<ShowDocumentParams | ResponseError<ShowDocumentResult>> => {
-            try {
-                const uri = vscode.Uri.parse(params.uri)
-                const doc = await vscode.workspace.openTextDocument(uri)
-                await vscode.window.showTextDocument(doc, { preview: false })
-                return params
-            } catch (e) {
-                return new ResponseError(
-                    LSPErrorCodes.RequestFailed,
-                    `Failed to open document: ${(e as Error).message}`
-                )
-            }
-        }
-    )
 
     languageClient.onNotification(contextCommandsNotificationType.method, (params: ContextCommandParams) => {
         void provider.webview?.postMessage({
